requirements.txt
altair==5.0.1
altgraph==0.17.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==5.1.1
gitdb==4.0.10
GitPython==3.1.31
google-auth
google-cloud-bigquery
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
jsonschema==4.17.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.24.3
openai
packaging==23.1
pandas==2.0.2
pandasai
pandas_gbq
pefile==2023.2.7
Pillow==9.5.0
protobuf==4.23.2
pyarrow==12.0.1
pydeck==0.8.1b0
Pygments==2.15.1
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
Pympler==1.0.1
pyrsistent==0.19.3
PySimpleGUI==4.60.5
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32-ctypes==0.2.0
requests==2.31.0
requests_oauthlib
rich==13.4.2
six==1.16.0
smmap==5.0.0
streamlit==1.23.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
torch
tornado==6.3.2
transformers
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
Unidecode==1.3.6
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
zipp==3.15.0